requirements.txt
alembic==1.4.2
astroid==2.4.2
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
colorama==0.4.3
easydict==1.9
firebase-admin==4.1.0
Flask==1.1.2
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
flask-serialize==1.1.9
Flask-SQLAlchemy==2.4.1
GeoAlchemy2==0.8.3
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.1
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.28.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
gunicorn==20.0.4
httplib2==0.17.3
idna==2.9
isort==5.5.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.8.0
mccabe==0.6.1
msgpack==1.0.0
protobuf==3.11.3
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygeohash==1.2.0
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.1
requests==2.23.0
rsa==4.0
Shapely==1.7.1
six==1.14.0
SQLAlchemy==1.3.16
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.9
validate-email==1.3
Werkzeug==1.0.1
wrapt==1.12.1
